[IMP] survey: add optional encapsulation layout when sending survey emails
This commit adds possibility to add a notification layout when sending
an email in survey using their specific composer. Purpose of this task is to
allow to send some technical template-based emails using the default
notification layouts defined in mail. It allows to avoid redefining the whole
layout in the template itself. This way email body content is simpler.

This commit is linked to task ID 1843224 (and 1868112) and to PR #25313
(and #25889).