[IMP] portal: always send creation / connection email to new portal users
Currently it is embedded in a strange condition about reactivating an
existing user or adding the portal group. Let us send the connection
email each time as it contains basic information on the account and is
useful for the newly created or activated user.

This commit is linked to task ID 1843361 (and 1868112) and to PR #25299
(and #25889).